requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
asyncpg
pydantic
pydantic-settings
python-jose
requests
stripe
